refactor: add new "Request" request and add proper closening of client
1- New "Request" struct is created to store the information of the
client request, along with several functions that iterates over the
request to do so.

2.- Added proper implementations of the "free()" function to free the
memory once the response has been aborted due to not found resources.

3.- A prototype of a Not Found 404 was added into the "main.c" source
file. Closing the client socket if not response was given provoked the
client to automatically send several ones to search for that response,
so a 404 is given so the client won't make those requests.